backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
dj-database-url
gunicorn
whitenoise
coverage
Pillow
django-storages[google]
google-cloud-storage
